Extract download logic for easier re-use
