+++ corehq/apps/hqadmin/utils.py
@@ -1,17 +1,20 @@
 import os
+import uuid
 from importlib import import_module
 from itertools import groupby
 
 from django.conf import settings
 from django.contrib.auth import SESSION_KEY, get_user_model
-from django.utils.safestring import mark_safe
 from django.utils.html import format_html
+from django.utils.safestring import mark_safe
 
 import requests
 
 from pillowtop.utils import force_seq_int
 
 from corehq.apps.hqadmin.models import HistoricalPillowCheckpoint
+from corehq.blobs import CODES, get_blob_db
+from corehq.util.view_utils import reverse
 
 EPSILON = 10000000
 
@@ -175,3 +178,19 @@ def unset_password(user):
     # 128 bits / 8 (bits/byte) = 16 bytes
     random_key = os.urandom(16).hex()
     user.set_password(random_key)
+
+
+def get_download_url(content, name, content_type=None, timeout=24 * 60):
+    """Upload file to blob storage for subsequent download"""
+    unique_id = str(uuid.uuid4())
+    get_blob_db().put(
+        content,
+        domain=unique_id,
+        parent_id=unique_id,
+        type_code=CODES.tempfile,
+        key=unique_id,
+        name=name,
+        content_type=content_type,  # optional
+        timeout=timeout,  # minutes
+    )
+    return reverse('download_blob', params={'key': unique_id}, absolute=True)
